+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Added support for sending MMS messages via Twilio with `media_urls` option.
 - Moved to TravisCI
 - Added support for using Twilio Copilot which depends on having a `messaging_service_sid`.
+- Fix defaults inheritance (see issue #11)